Use enhanced pin example working
